MANUAL: renames files and component to selection-tree
