+++ apps/core-ui-e2e/src/e2e/selection-tree.cy.ts
