+++ libs/core/ui/src/index.ts
@@ -2,7 +2,7 @@ export * from './lib/animations/animated-component.class';
 export * from './lib/autofocus/autofocus.directive';
 export * from './lib/button/button.component';
 export * from './lib/checkbox/checkbox.component';
-export * from './lib/checkbox-tree/checkbox-tree.component';
+export * from './lib/selection-tree/selection-tree.component';
 export * from './lib/counted-checkbox-tree/counted-checkbox-tree.component';
 export * from './lib/external-link/external-link.directive';
 export * from './lib/forms/form.type';
