+++ libs/core/ui/src/lib/selection-tree/selection-tree.component.scss
